AgenticAI/requirements.txt
phidata
python-dotenv
fastapi
uvicorn
yfinance
duckduckgo-search
packaging
groq
openai
